v1/pyproject.toml
[project]
name = "llm-finetune-pipeline"
version = "0.1.0"

description = "morpheme-segmentation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.4.0",
    "transformers>=4.46.0",
    "datasets>=2.21.0",
    "peft>=0.13.0",
    "trl>=0.11.0",
    "bitsandbytes>=0.44.0",
    "ruamel-yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "ty",
]

[tool.ruff]
line-length = 120
target-version = "py313"

lint.select = ["E", "F", "W"]
lint.ignore = []
lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ty]
pythonVersion = ">=3.13"
strict = false



